Finished setting up and styling delete question popup
