+++ src/components/MonitorLog/GeneralInfoTab.jsx
@@ -37,6 +37,7 @@ import { React, useState, useEffect } from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import { Controller, useFormContext } from 'react-hook-form';
+import NonStaticQuestion from './NonStaticQuestion';
 import './GeneralInfoTab.css';
 import { OCHBackend } from '../../common/utils';
 
@@ -45,37 +46,37 @@ function GeneralInfoTab({ assignedSegments, monitorPartners, isDisabled, showHea
   // commented for prettier
   // const [isLoading, setIsLoading] = useState(true);
   const { control, register } = useFormContext();
-  const editQuestionModal = useDisclosure();
+  // const editQuestionModal = useDisclosure();
 
   // states for editQuestion modal
-  const [newTitle, setNewTitle] = useState();
-  const [newFieldType, setNewFieldType] = useState();
-  const [newTooltip, setNewTooltip] = useState();
+  // const [newTitle, setNewTitle] = useState();
+  // const [newFieldType, setNewFieldType] = useState();
+  // const [newTooltip, setNewTooltip] = useState();
   // const [fieldToEdit, setFieldToEdit] = useState();
   const [idOfFieldBeingEdited, setIdOfFieldBeingEdited] = useState();
 
-  const updateQuestion = async () => {
-    console.log(`updateQuestion called with fieldId: ${idOfFieldBeingEdited}`);
-    console.log(`fieldBody: ${newTitle}, ${newFieldType}, ${newTooltip}`);
-    await OCHBackend.put('/forms/update/field', {
-      type: 'general',
-      fieldId: idOfFieldBeingEdited,
-      fieldBody: {
-        title: newTitle,
-        fieldType: newFieldType,
-        tooltip: newTooltip,
-      },
-    });
-    editQuestionModal.onClose();
-  };
+  // const updateQuestion = async () => {
+  //   console.log(`updateQuestion called with fieldId: ${idOfFieldBeingEdited}`);
+  //   console.log(`fieldBody: ${newTitle}, ${newFieldType}, ${newTooltip}`);
+  //   await OCHBackend.put('/forms/update/field', {
+  //     type: 'general',
+  //     fieldId: idOfFieldBeingEdited,
+  //     fieldBody: {
+  //       title: newTitle,
+  //       fieldType: newFieldType,
+  //       tooltip: newTooltip,
+  //     },
+  //   });
+  //   editQuestionModal.onClose();
+  // };
 
-  const deleteQuestion = async () => {
-    await OCHBackend.delete('/forms/delete/field', {
-      formType: 'general',
-      fieldId: idOfFieldBeingEdited,
-    });
-    editQuestionModal.onClose();
-  };
+  // const deleteQuestion = async () => {
+  //   await OCHBackend.delete('/forms/delete/field', {
+  //     formType: 'general',
+  //     fieldId: idOfFieldBeingEdited,
+  //   });
+  //   editQuestionModal.onClose();
+  // };
 
   useEffect(async () => {
     const newQuestions = await OCHBackend.get(`/forms/general`);
@@ -279,48 +280,49 @@ function GeneralInfoTab({ assignedSegments, monitorPartners, isDisabled, showHea
           </GridItem>
           {additionalQuestions.map(question => {
             return (
-              <Box
-                w="255px"
-                h="133px"
-                borderRadius="6px"
-                key={question.title}
-                _hover={isTemplate ? { bgColor: 'rgba(43, 192, 227, 0.25)' } : null}
-                onClick={
-                  isTemplate
-                    ? () => {
-                        editQuestionModal.onOpen();
-                        setNewTitle(question.title);
-                        setNewFieldType(question.fieldType);
-                        setNewTooltip(question.tooltip);
-                        setIdOfFieldBeingEdited(question._id);
-                        console.log(`idOfFieldBeingEdited: ${idOfFieldBeingEdited}`);
-                        console.log(additionalQuestions);
-                      }
-                    : null
-                }
-                px="10px"
-                py="10px"
-              >
-                <GridItem key={question.title} colSpan={1} rowSpan={1}>
-                  <VStack spacing="8px" align="left">
-                    <Text fontWeight="500" fontSize="md">
-                      {question.title}
-                    </Text>
-                    {question.fieldType === 'TEXT' ? (
-                      <Input type="text" />
-                    ) : (
-                      <NumberInput allowMouseWheel>
-                        <NumberInputField />
-                        <NumberInputStepper>
-                          <NumberIncrementStepper />
-                          <NumberDecrementStepper />
-                        </NumberInputStepper>
-                      </NumberInput>
-                    )}
-                    {isTemplate && <Text color="#718096">Non-Static</Text>}
-                  </VStack>
-                </GridItem>
-              </Box>
+              <NonStaticQuestion key={question.title} question={question} />
+              // <Box
+              //   w="255px"
+              //   h="133px"
+              //   borderRadius="6px"
+              //   key={question.title}
+              //   _hover={isTemplate ? { bgColor: 'rgba(43, 192, 227, 0.25)' } : null}
+              //   onClick={
+              //     isTemplate
+              //       ? () => {
+              //           editQuestionModal.onOpen();
+              //           setNewTitle(question.title);
+              //           setNewFieldType(question.fieldType);
+              //           setNewTooltip(question.tooltip);
+              //           setIdOfFieldBeingEdited(question._id);
+              //           console.log(`idOfFieldBeingEdited: ${idOfFieldBeingEdited}`);
+              //           console.log(additionalQuestions);
+              //         }
+              //       : null
+              //   }
+              //   px="10px"
+              //   py="10px"
+              // >
+              //   <GridItem key={question.title} colSpan={1} rowSpan={1}>
+              //     <VStack spacing="8px" align="left">
+              //       <Text fontWeight="500" fontSize="md">
+              //         {question.title}
+              //       </Text>
+              //       {question.fieldType === 'TEXT' ? (
+              //         <Input type="text" />
+              //       ) : (
+              //         <NumberInput allowMouseWheel>
+              //           <NumberInputField />
+              //           <NumberInputStepper>
+              //             <NumberIncrementStepper />
+              //             <NumberDecrementStepper />
+              //           </NumberInputStepper>
+              //         </NumberInput>
+              //       )}
+              //       {isTemplate && <Text color="#718096">Non-Static</Text>}
+              //     </VStack>
+              //   </GridItem>
+              // </Box>
             );
           })}
         </SimpleGrid>
@@ -354,7 +356,7 @@ function GeneralInfoTab({ assignedSegments, monitorPartners, isDisabled, showHea
       </VStack>
 
       {/* EDIT QUESTION MODAL STARTS HERE */}
-      <Modal
+      {/* <Modal
         w="460px"
         h="562px"
         bgColor="rgba(253, 253, 253, 1)"
@@ -429,7 +431,7 @@ function GeneralInfoTab({ assignedSegments, monitorPartners, isDisabled, showHea
             </Box>
           </ModalBody>
         </ModalContent>
-      </Modal>
+      </Modal> */}
     </div>
   );
 }
